backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.0
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
environs==11.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
loguru==0.7.2
marshmallow==3.23.1
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.20.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
win32-setctime==1.1.0
